tests/conftest.py
import pytest

from app import create_app


@pytest.fixture
def client():
    app = create_app({"TESTING": True})
    with app.test_client() as client:
        yield client

def app():
    app = create_app({"TESTING": True})
    return app