requirements.txt
python-dotenv
sqlalchemy
scikit-learn
fastapi
flask
libsass
pandas
uvicorn
python-multipart
jinja2
pandas
seaborn
flask-cors
openpyxl
pytest-flask